[SPARK-49425][CONNECT][SQL] Create a shared DataFrameWriter
### What changes were proposed in this pull request?
This PR adds a shared interface for DataFrameWriter.

### Why are the changes needed?
We are creating a shared interface for Classic and Connect Scala SQL API. This class is part of that effort.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47922 from hvanhovell/SPARK-49425.

Authored-by: Herman van Hovell <herman@databricks.com>
